Implement agenda edit modals and refresh documentation
