+++ docs/remaining-work.md
@@ -4,31 +4,32 @@
 
 | 영역 | 미완료 항목 | 영향 | 다음 조치 |
 | --- | --- | --- | --- |
-| 편집 플로우 | 상세 시트의 `편집` 버튼이 내비게이션과 연결되지 않음 | 사용자가 기존 일정/할 일을 수정할 방법이 없음 | 전용 편집 화면/다이얼로그를 만들고 `AgendaRoute` 콜백을 실제 내비게이션과 연결 |
 | 리마인더 구성 | 빠른 추가 시 리마인더 오프셋·스누즈를 설정할 수 없음 | 요구사항상의 사용자 정의 리마인더 경험 부족 | 빠른 추가 및 정식 편집 플로우에 리마인더 입력 필드를 추가하고 `ReminderOrchestrator`에 전달 |
 | 현지화 | Compose 내부에 한국어 문자열이 하드코딩되어 있음 | 다국어/접근성 대응 불가 | 문자열을 `strings.xml`로 추출하고 번역 키를 준비 |
 | UI 구조 | Agenda 화면에 유사한 카드/리스트 컴포저블이 중복 선언됨 | 유지보수성·스크린 리더 탐색성 저하 | 공통 컴포넌트를 정리하고 의미 있는 heading/role 속성을 재검토 |
 | 동기화 | `docs/sync-conflict-strategy.md`에 정의된 클라우드 동기화 정책 미구현 | 멀티 디바이스/오프라인 사용 시 데이터 손실 우려 | 서버/동기화 계층을 도입하고 충돌 해결 정책을 실제 코드로 이전 |
 
+## 완료된 항목
+
+### 편집 플로우 보완 (완료)
+- 상세 시트에서 `편집`을 누르면 상위 `CalendarApp`이 전용 편집 모달 시트를 열어 기존 일정·할 일을 수정할 수 있습니다.【F:app/src/main/kotlin/com/example/calendar/ui/CalendarApp.kt†L20-L121】【F:app/src/main/kotlin/com/example/calendar/ui/agenda/EditSheets.kt†L1-L170】
+- `AgendaRoute`는 편집 버튼을 누를 때 상세 시트를 닫고 콜백을 호출해 상위 내비게이션과 자연스럽게 연동합니다.【F:app/src/main/kotlin/com/example/calendar/ui/agenda/AgendaScreen.kt†L264-L310】
+
 ## 세부 항목
 
-### 1. 편집 플로우 보완
-- `TaskDetailSheet`와 `EventDetailSheet`의 `편집` 버튼은 `onTaskEdit`/`onEventEdit` 콜백만 호출하며, 기본 구현은 비어 있습니다.【F:app/src/main/kotlin/com/example/calendar/ui/agenda/AgendaScreen.kt†L1404-L1477】
-- 실제 편집 화면 또는 다이얼로그를 정의하고, `AgendaRoute` 상위에서 내비게이션을 제공해야 합니다.【F:app/src/main/kotlin/com/example/calendar/ui/agenda/AgendaScreen.kt†L103-L129】
-
-### 2. 리마인더 구성 UI 추가
-- 빠른 추가 시트는 제목·메모·시간만 입력받아 리마인더 설정을 변경할 수 없습니다.【F:app/src/main/kotlin/com/example/calendar/ui/agenda/AgendaScreen.kt†L1479-L1662】
+### 1. 리마인더 구성 UI 추가
+- 빠른 추가 시트는 제목·메모·시간만 입력받아 리마인더 설정을 변경할 수 없습니다.【F:app/src/main/kotlin/com/example/calendar/ui/agenda/AgendaScreen.kt†L1512-L1754】
 - `AgendaViewModel`도 기본 리마인더 이외에 오프셋 목록을 처리하지 않습니다.【F:app/src/main/kotlin/com/example/calendar/ui/AgendaViewModel.kt†L57-L152】
 - 사용자가 선택한 오프셋을 `ReminderPayload`로 변환해 `ReminderOrchestrator`에 전달하는 로직이 필요합니다.【F:app/src/main/kotlin/com/example/calendar/reminder/ReminderOrchestrator.kt†L16-L91】
 
-### 3. 현지화 리소스 정비
+### 2. 현지화 리소스 정비
 - 대다수 Compose 텍스트가 하드코딩되어 있고, `strings.xml`에는 앱 이름과 알림 관련 문자열만 존재합니다.【F:app/src/main/kotlin/com/example/calendar/ui/agenda/AgendaScreen.kt†L707-L1165】【F:app/src/main/res/values/strings.xml†L1-L8】
 - 문자열을 리소스로 추출하고, 접근성 설명(`contentDescription`, `stateDescription`)도 Locale 대응이 가능하도록 개선해야 합니다.【F:app/src/main/kotlin/com/example/calendar/ui/agenda/AgendaScreen.kt†L914-L1100】
 
-### 4. Agenda UI 구조 개선
+### 3. Agenda UI 구조 개선
 - `AgendaScreen`에는 리스트/요약 컴포저블이 유사한 형태로 중복 선언되어 있어 유지보수가 어렵습니다.【F:app/src/main/kotlin/com/example/calendar/ui/agenda/AgendaScreen.kt†L807-L1406】
 - 기존 Compose UI 테스트는 기본 렌더링만 검증하므로, 접근성 포커스·스크린 리더 플로우를 확인하는 시나리오가 추가로 필요합니다.【F:app/src/androidTest/kotlin/com/example/calendar/ui/agenda/AgendaScreenTest.kt†L19-L70】
 
-### 5. 클라우드 동기화 구현
+### 4. 클라우드 동기화 구현
 - 클라우드 동기화 및 충돌 해결 전략은 문서로만 존재하며 실제 코드에서는 동작하지 않습니다.【F:docs/sync-conflict-strategy.md†L1-L33】
 - Room과 원격 소스를 연결하는 동기화 서비스, 충돌 히스토리 저장, 사용자 알림 흐름을 구현해야 합니다.
